COM-1648 review types organization
