+++ src/api/activities.ts
@@ -1,6 +1,6 @@
 import { alenviAxios } from './ressources/alenviAxios';
 import getEnvVars from '../../environment';
-import { ActivityWithCardsType } from '../types/CourseType';
+import { ActivityWithCardsType } from '../types/ActivityTypes';
 
 export default {
   getActivity: async (activityId): Promise<ActivityWithCardsType> => {
