+++ src/components/About/index.tsx
@@ -12,7 +12,7 @@ import NiPrimaryButton from '../../components/form/PrimaryButton';
 import FeatherButton from '../../components/icons/FeatherButton';
 import CoursesActions from '../../store/courses/actions';
 import { ActionWithoutPayloadType } from '../../types/store/StoreType';
-import { ProgramType } from '../../types/CourseType';
+import { ProgramType } from '../../types/CourseTypes';
 import FooterGradient from '../design/FooterGradient';
 
 type AboutProps = {
