+++ src/components/activities/ActivityIcon/index.tsx
@@ -5,7 +5,7 @@ import LessonIcon from '../../../../assets/icons/LessonIcon';
 import QuizIcon from '../../../../assets/icons/QuizIcon';
 import SharingExperienceIcon from '../../../../assets/icons/SharingExperienceIcon';
 import commonStyle from '../../../styles/common';
-import { ActivityType } from '../../../types/CourseType';
+import { ActivityType } from '../../../types/ActivityTypes';
 import { SHARING_EXPERIENCE, LESSON, QUIZ, VIDEO } from '../../../core/data/constants';
 import Shadow from '../../design/Shadow';
 import styles from './styles';
