+++ src/store/cards/actions.ts
@@ -4,7 +4,6 @@ import {
   AddQuestionnaireAnswerType,
   RemoveQuestionnaireAnswerType,
   SetQuestionnaireAnswersListType,
-  QuestionnaireAnswerType,
   SetExitConfirmationModalType,
   SET_CARDS,
   RESET_CARD_REDUCER,
@@ -17,17 +16,18 @@ import {
 } from '../../types/store/CardStoreType';
 import { CardType } from '../../types/CardType';
 import { ActionWithoutPayloadType } from '../../types/store/StoreType';
+import { QuestionnaireAnswersType } from '../../types/ActivityTypes';
 
 const resetCardReducer = (): ActionWithoutPayloadType => ({ type: RESET_CARD_REDUCER });
 
-const setCards = (cards: Array<CardType>): SetCardsType => ({ type: SET_CARDS, payload: cards });
+const setCards = (cards: CardType[]): SetCardsType => ({ type: SET_CARDS, payload: cards });
 
 const setCardIndex = (index: number): SetCardIndexType => ({ type: SET_CARD_INDEX, payload: index });
 
 const setQuestionnaireAnswersList =
-(questionnaireAnswersList: Array<QuestionnaireAnswerType>): SetQuestionnaireAnswersListType =>
+(questionnaireAnswersList: QuestionnaireAnswersType[]): SetQuestionnaireAnswersListType =>
   ({ type: SET_QUESTIONNAIRE_ANSWERS_LIST, payload: questionnaireAnswersList });
-const addQuestionnaireAnswer = (questionnaireAnswer: QuestionnaireAnswerType): AddQuestionnaireAnswerType =>
+const addQuestionnaireAnswer = (questionnaireAnswer: QuestionnaireAnswersType): AddQuestionnaireAnswerType =>
   ({ type: ADD_QUESTIONNAIRE_ANSWER, payload: questionnaireAnswer });
 const removeQuestionnaireAnswer = (card: string): RemoveQuestionnaireAnswerType =>
   ({ type: REMOVE_QUESTIONNAIRE_ANSWER, payload: card });
