+++ src/store/cards/selectors.ts
@@ -1,9 +1,9 @@
 import { OPEN_QUESTION, QUESTION_ANSWER, SURVEY, TRANSITION } from '../../core/data/constants';
-import { QuestionnaireAnswerType } from '../../types/store/CardStoreType';
+import { QuestionnaireAnswersType } from '../../types/ActivityTypes';
 
 const getCard = state => state.cards.cards[state.cards.cardIndex];
 
-const getQuestionnaireAnswer = (state): QuestionnaireAnswerType | null => {
+const getQuestionnaireAnswer = (state): QuestionnaireAnswersType | null => {
   const card = getCard(state);
   if (!card || (![SURVEY, OPEN_QUESTION, QUESTION_ANSWER].includes(card.template))) return null;
   return state.cards.questionnaireAnswersList.find(qa => qa.card === card._id) || null;
