+++ src/types/ActivityHistoryType.ts
@@ -1,10 +1,10 @@
-import { QuestionnaireAnswerType } from './store/CardStoreType';
+import { QuestionnaireAnswersType } from './ActivityTypes';
 
 export interface ActivityHistoryType {
   _id: string,
   user: string,
   activity: string,
-  questionnaireAnswersList?: Array<QuestionnaireAnswerType>,
+  questionnaireAnswersList?: QuestionnaireAnswersType[],
   date: Date,
   score: number,
 }
