+++ src/types/store/CardStoreType.ts
@@ -1,4 +1,5 @@
 import { CardType } from '../CardType';
+import { QuestionnaireAnswersType } from '../ActivityTypes';
 
 export const RESET_CARD_REDUCER = 'RESET_CARD_REDUCER';
 export const SET_CARDS = 'SET_CARDS';
@@ -13,7 +14,7 @@ export const SET_EXIT_CONFIRMATION_MODAL = 'SET_EXIT_CONFIRMATION_MODAL';
 export interface CardStateType {
   cards: Array<CardType>,
   cardIndex: number | null,
-  questionnaireAnswersList: Array<QuestionnaireAnswerType>,
+  questionnaireAnswersList: QuestionnaireAnswersType[],
   score: number,
   exitConfirmationModal: boolean,
 }
@@ -25,15 +26,13 @@ export interface SetCardsType { type: typeof SET_CARDS, payload: Array<CardType>
 
 export interface SetCardIndexType { type: typeof SET_CARD_INDEX, payload: number }
 
-export interface QuestionnaireAnswerType { _id?: string, card: string, answerList: Array<string> }
-
-export interface AddQuestionnaireAnswerType { type: typeof ADD_QUESTIONNAIRE_ANSWER, payload: QuestionnaireAnswerType }
+export interface AddQuestionnaireAnswerType { type: typeof ADD_QUESTIONNAIRE_ANSWER, payload: QuestionnaireAnswersType }
 
 export interface RemoveQuestionnaireAnswerType { type: typeof REMOVE_QUESTIONNAIRE_ANSWER, payload: string }
 
 export interface SetQuestionnaireAnswersListType {
   type: typeof SET_QUESTIONNAIRE_ANSWERS_LIST,
-  payload: Array<QuestionnaireAnswerType>,
+  payload: QuestionnaireAnswersType[],
 }
 
 export interface IncGoodAnswersCountType { type: typeof INC_GOOD_ANSWERS_COUNT }
